+++ src/web-auth/username-password.js
@@ -11,7 +11,7 @@ import paramsArray from '../helper/param-constants';
 function UsernamePassword(options) {
   this.baseOptions = options;
   this.request = new RequestBuilder(options);
-  this.transactionManager = new TransactionManager(this.baseOptions.transaction);
+  this.transactionManager = new TransactionManager(this.baseOptions);
 }
 
 UsernamePassword.prototype.login = function(options, cb) {